refactor(settings): separate configuration responsibilities
Move settings-file loading and validation out of the keymap module. Keep keymap parsing focused on bindings and let the theme layer resolve supported theme names and defaults.\n\nAssisted-by: Codex GPT-5
